refactor code based on feedback
