+++ index.html
@@ -231,7 +231,7 @@
 <script type="text/javascript" src="lib/js/bootstrap.min.js"></script>
 <script type="text/javascript" src="lib/js/angular.min.js"></script>
 <script type="text/javascript" src="src/js/app.js"></script>
-<script type="text/javascript" src="src/js/invertedIndex.js"></script>
+<script type="text/javascript" src="src/js/InvertedIndex.js"></script>
 <script type="text/javascript" src="/lib/js/dir-pagination.js"></script>
 
 </html>
